+++ inlang/source-code/paraglide/paraglide-next/src/cli/commands/init.ts
@@ -1,29 +1,17 @@
-import { Repository, findRepoRoot, openRepository } from "@lix-js/client"
-import type { CliStep } from "../utils"
+import { findRepoRoot, openRepository } from "@lix-js/client"
 import { Command } from "commander"
 import nodeFsPromises from "node:fs/promises"
-import { NodeishFilesystem } from "@lix-js/fs"
 import path from "node:path"
 import { Logger } from "@inlang/paraglide-js/internal"
 import { Steps, cli as ParaglideCli } from "@inlang/paraglide-js/internal/cli"
-import consola from "consola"
-import { SetUpI18nRoutingFlow } from "../flows/set-up-i18n-routing"
+import { scanNextJSProject } from "../flows/scan-next-project"
+import { updateNextConfig } from "../flows/updateNextConfig"
+import { createI18nFile } from "../flows/app-router-setup/createI18nFile"
+import { createMiddlewareFile } from "../flows/app-router-setup/createMiddlewareFile"
+import { addLanguageProvider } from "../flows/app-router-setup/addLanguageProvider"
+import { maybeMigrateI18nRouting } from "../flows/app-router-setup/maybeMigrateI18nRouting"
 
-type NextConfigFile = {
-	path: string
-	type: "cjs" | "esm"
-}
-type NextJSProject = {
-	configFile: NextConfigFile
-	/** The absolute path to the src folder (if present), otherwise the current working directory  */
-	srcRoot: string
-	/** If the project uses TypeScript */
-	typescript: boolean
-	/** Which router the Next App uses */
-	router: "app" | "pages"
-}
-
-type Outdir = {
+export type Outdir = {
 	/**
 	 * The absolute path to the outdir
 	 * @example
@@ -79,13 +67,18 @@ export const InitCommand = new Command()
 			}),
 		})(ctx4)
 		const ctx6 = await Steps.maybeChangeTsConfig(ctx5)
-		const ctx7 = await createI18nFile(ctx6)
-		const ctx8 = await createMiddlewareFile(ctx7)
-		const ctx9 = await updateNextConfig(ctx8)
-		const ctx10 = await addLanguageProvider(ctx9)
-		const ctx11 = await maybeMigrateI18nRouting(ctx10)
+
+		// addRouter Setup
+		if (ctx6.nextProject.router == "pages") {
+			await createI18nFile(ctx6)
+			await createMiddlewareFile(ctx6)
+			await addLanguageProvider(ctx6)
+			await updateNextConfig(ctx6)
+			await maybeMigrateI18nRouting(ctx6)
+		}
+
 		try {
-			await Steps.runCompiler({ ...ctx11, outdir: ctx11.outdir.path })
+			await Steps.runCompiler({ ...ctx6, outdir: ctx6.outdir.path })
 		} catch (e) {
 			//silently ignore
 		}
@@ -96,388 +89,3 @@ Learn more about Paraglide and Paraglide-Next at:
 https://inlang.com/m/osslbuzt/paraglide-next-i18n
 `)
 	})
-
-const scanNextJSProject: CliStep<
-	{
-		repo: Repository
-		logger: Logger
-	},
-	{
-		nextProject: NextJSProject
-		/** The absolute path to the package.json file  */
-		packageJsonPath: string
-	}
-> = async (ctx) => {
-	const packageJsonPath = await findPackageJson(ctx.repo.nodeishFs, process.cwd())
-	if (!packageJsonPath) {
-		ctx.logger.error(`Could not find package.json. Rerun this command inside a NextJS project.`)
-		process.exit(1)
-	}
-
-	const configFile = await findNextConfig(ctx.repo.nodeishFs, process.cwd())
-	if (!configFile) {
-		ctx.logger.error(`Could not find Next Config File. Rerun this command inside a NextJS project.`)
-		process.exit(1)
-	}
-
-	// try to read tsconfig or JSConfig
-	const tsConfigPath = path.join(process.cwd(), "tsconfig.json")
-	const jsConfigPath = path.join(process.cwd(), "jsconfig.json")
-
-	let typescript = false
-	try {
-		const stat = await ctx.repo.nodeishFs.stat(tsConfigPath)
-		if (stat.isFile()) {
-			typescript = true
-		}
-	} catch {
-		//silently ignore
-	}
-
-	try {
-		const stat = await ctx.repo.nodeishFs.stat(jsConfigPath)
-		if (stat.isFile()) {
-			typescript = false
-		}
-	} catch {
-		//silently ignore
-	}
-
-	// if the ./src directory exists -> srcRoot = ./src
-	// otherwise -> srcRoot  = .
-
-	let srcRoot
-	try {
-		const stat = await ctx.repo.nodeishFs.stat(path.resolve(process.cwd(), "src"))
-		if (!stat.isDirectory()) throw Error()
-		srcRoot = path.resolve(process.cwd(), "src")
-	} catch {
-		srcRoot = process.cwd()
-	}
-
-	let router: "app" | "pages" | undefined = undefined
-	const expectedAppFolderPath = path.join(srcRoot, "app")
-	const expectedPagesFolderPath = path.join(srcRoot, "pages")
-
-	try {
-		const stat = await ctx.repo.nodeishFs.stat(expectedPagesFolderPath)
-		if (!stat.isDirectory()) throw new Error()
-		router = "pages"
-	} catch {
-		//silently ignore
-	}
-
-	try {
-		const stat = await ctx.repo.nodeishFs.stat(expectedAppFolderPath)
-		if (!stat.isDirectory()) throw new Error()
-		router = "app"
-	} catch {
-		//silently ignore
-	}
-
-	if (!router) {
-		ctx.logger.error(
-			"Failed to determine the NextJS router. Please make sure either the app or pages folders exists in your project"
-		)
-		process.exit(1)
-	}
-
-	return {
-		...ctx,
-		packageJsonPath,
-		nextProject: {
-			srcRoot,
-			configFile,
-
-			typescript,
-			router,
-		},
-	}
-}
-
-/**
- * Attempts to find the next.config.js or next.config.mjs file in the current working directory.
- */
-async function findNextConfig(
-	fs: NodeishFilesystem,
-	cwd: string
-): Promise<NextConfigFile | undefined> {
-	const possibleNextConfigPaths = ["./next.config.js", "./next.config.mjs"].map(
-		(possibleRelativePath) => path.resolve(cwd, possibleRelativePath)
-	)
-
-	for (const possibleNextConfigPath of possibleNextConfigPaths) {
-		try {
-			const stat = await fs.stat(possibleNextConfigPath)
-			if (!stat.isFile()) continue
-
-			return {
-				path: possibleNextConfigPath,
-				type: possibleNextConfigPath.endsWith(".mjs") ? "esm" : "cjs",
-			}
-		} catch {
-			continue
-		}
-	}
-
-	return undefined
-}
-
-/**
- * Attempts to find the package.json file in the current working directory.
- */
-async function findPackageJson(fs: NodeishFilesystem, cwd: string): Promise<string | undefined> {
-	const potentialPackageJsonPath = path.resolve(cwd, "package.json")
-	try {
-		const stat = await fs.stat(potentialPackageJsonPath)
-		if (!stat.isFile()) {
-			return undefined
-		}
-		return potentialPackageJsonPath
-	} catch {
-		return undefined
-	}
-}
-
-const createI18nFile: CliStep<
-	{
-		repo: Repository
-		nextProject: NextJSProject
-	},
-	unknown
-> = async (ctx) => {
-	const i18nFilePath = path.join(
-		ctx.nextProject.srcRoot,
-		ctx.nextProject.typescript ? "lib/i18n.ts" : "lib/i18n.js"
-	)
-
-	const TypescriptTemplate = `// file generated by the Paraglide-Next init command
-import { createI18n } from "@inlang/paraglide-next"
-import type { AvailableLanguageTag } from "@/paraglide/runtime"
-
-export const { Link, middleware, useRouter, usePathname, redirect, permanentRedirect, localizePath } 
-	= createI18n<AvailableLanguageTag>();
-`
-
-	const JavascriptTemplate = `// file generated by the Paraglide-Next init command
-import { createI18n } from "@inlang/paraglide-next"
-
-export const { Link, middleware, useRouter, usePathname, redirect, permanentRedirect, localizePath }
-	= createI18n();
-`
-
-	const file = ctx.nextProject.typescript ? TypescriptTemplate : JavascriptTemplate
-
-	//create the folder if it doesn't exist
-	await ctx.repo.nodeishFs.mkdir(path.dirname(i18nFilePath), { recursive: true })
-	await ctx.repo.nodeishFs.writeFile(i18nFilePath, file)
-	return ctx
-}
-
-const createMiddlewareFile: CliStep<
-	{ repo: Repository; logger: Logger; nextProject: NextJSProject },
-	unknown
-> = async (ctx) => {
-	const middlewareFilePath = path.join(
-		ctx.nextProject.srcRoot,
-		ctx.nextProject.typescript ? "middleware.ts" : "middleware.js"
-	)
-
-	//check if the middleware file already exists
-	let alreadyExists: boolean
-	try {
-		await ctx.repo.nodeishFs.stat(middlewareFilePath)
-		alreadyExists = true
-	} catch {
-		//if the middleware file doesn't exist, create it
-		alreadyExists = false
-	}
-
-	if (alreadyExists) {
-		ctx.logger.warn(
-			"Skipping creating the `middleware.ts` file as it already exists. Please manually add the middleware exported from @/lib/i18n"
-		)
-		return ctx
-	}
-
-	const file = `// file generated by the Paraglide-Next init command
-export { middleware } from "@/lib/i18n"
-
-export const config = {
-	matcher: [
-		/*
-		 * Match all request paths except for the ones starting with:
-		 * - api (API routes)
-		 * - _next/static (static files)
-		 * - _next/image (image optimization files)
-		 * - favicon.ico (favicon file)
-		 */
-		"/((?!api|_next/static|_next/image|favicon.ico).*)",
-	],
-}
-`
-
-	//create the folder if it doesn't exist
-	await ctx.repo.nodeishFs.mkdir(path.dirname(middlewareFilePath), { recursive: true })
-	await ctx.repo.nodeishFs.writeFile(middlewareFilePath, file)
-	return ctx
-}
-
-const updateNextConfig: CliStep<
-	{
-		logger: Logger
-		repo: Repository
-		outdir: Outdir
-		projectPath: string
-		nextProject: NextJSProject
-	},
-	unknown
-> = async (ctx) => {
-	//read the next.config.js file
-	let fileContent: string
-	try {
-		fileContent = await ctx.repo.nodeishFs.readFile(ctx.nextProject.configFile.path, {
-			encoding: "utf-8",
-		})
-	} catch (e) {
-		ctx.logger.error("Failed to read next config file at " + ctx.nextProject.configFile.path)
-		process.exit(1)
-	}
-
-	if (fileContent.includes("paraglide")) {
-		ctx.logger.warn(
-			"Skipping adding the paraglide plugin to `next.config.js` as it already seems to be added"
-		)
-		return ctx
-	}
-
-	//Add the import
-	const importStatement: string = {
-		esm: 'import { paraglide } from "@inlang/paraglide-next/plugin"',
-		cjs: 'const { paraglide } = require("@inlang/paraglide-next/plugin")',
-	}[ctx.nextProject.configFile.type]
-
-	fileContent = importStatement + "\n" + fileContent
-
-	const exportRegex = {
-		esm: /export\s+default\s+(?<configIdentifier>[a-zA-Z0-9]+)(?=\s|;)/gm,
-		cjs: /module.exports\s+=\s+(?<configIdentifier>[a-zA-Z0-9]+)(?=\s|;)/gm,
-	}[ctx.nextProject.configFile.type]
-
-	const match = exportRegex.exec(fileContent)
-	if (!match) {
-		ctx.logger.warn(
-			`Failed to find the export default statement in next.config.js
-You will have to add the paraglide plugin manually
-
-Learn how to do that in the documentation:
-https://inlang.com/m/osslbuzt/paraglide-next-i18n
-`
-		)
-	} else {
-		const exportDefault = match
-		const startIndex = exportDefault.index
-		const endIndex = startIndex + exportDefault[0].length
-		const configIdentifier = match.groups?.configIdentifier as string
-		const identifierStartIndex = endIndex - configIdentifier.length
-
-		const relativeOutdir = "./" + path.relative(process.cwd(), ctx.outdir.path)
-		const relativeProjectPath = "./" + path.relative(process.cwd(), ctx.projectPath)
-
-		const wrappedIdentifier = `paraglide({
-	paraglide: {
-		project: "${relativeProjectPath}",
-		outdir: "${relativeOutdir}"
-	},
-	...${configIdentifier}
-})`
-
-		//replace the wrapped identifier with the actual identifier
-		fileContent =
-			fileContent.slice(0, Math.max(0, identifierStartIndex)) +
-			wrappedIdentifier +
-			fileContent.slice(Math.max(0, endIndex))
-
-		ctx.logger.info("Added the paraglide plugin to next.config.js")
-		await ctx.repo.nodeishFs.writeFile(ctx.nextProject.configFile.path, fileContent)
-	}
-
-	return ctx
-}
-
-const addLanguageProvider: CliStep<
-	{
-		repo: Repository
-		logger: Logger
-		nextProject: NextJSProject
-	},
-	unknown
-> = async (ctx) => {
-	const layoutFilePath = path.join(
-		ctx.nextProject.srcRoot,
-		ctx.nextProject.typescript ? "app/layout.tsx" : "app/layout.js"
-	)
-	let layoutFileContent: string
-	try {
-		layoutFileContent = await ctx.repo.nodeishFs.readFile(layoutFilePath, { encoding: "utf-8" })
-	} catch (e) {
-		ctx.logger.warn(
-			"Failed to add the `<LanguageProvider>` to `app/layout.tsx`. You'll need to add it yourself"
-		)
-		return ctx
-	}
-
-	if (layoutFileContent.includes("LanguageProvider")) {
-		ctx.logger.warn(
-			"Skipping add ingthe `<LanguageProvider>` to `app/layout.tsx` as it already seems to be added"
-		)
-		return ctx
-	}
-
-	layoutFileContent = `import { LanguageProvider } from "@inlang/paraglide-next"\nimport { languageTag } from "@/paraglide/runtime.js"\n${layoutFileContent}`
-	layoutFileContent = layoutFileContent.replace('lang="en"', `lang={languageTag()}`)
-
-	//find the "<html" literal & it's indentation
-	const htmlIndex = layoutFileContent.indexOf("<html")
-	if (htmlIndex === -1) {
-		ctx.logger.warn(
-			"Failed to add the `<LanguageProvider>` to `app/layout.tsx`. You'll need to add it yourself"
-		)
-		return ctx
-	}
-
-	const indentationMatch = layoutFileContent.slice(0, htmlIndex).match(/\s+/)
-	const htmlIndentation = indentationMatch?.[indentationMatch.length - 1] || "    "
-
-	layoutFileContent = layoutFileContent.replace(
-		"<html",
-		`<LanguageProvider>\n${htmlIndentation}  <html`
-	)
-	layoutFileContent = layoutFileContent.replace(
-		"/html>",
-		`/html>\n${htmlIndentation}</LanguageProvider>`
-	)
-
-	await ctx.repo.nodeishFs.writeFile(layoutFilePath, layoutFileContent)
-	return ctx
-}
-
-const maybeMigrateI18nRouting: CliStep<
-	{
-		repo: Repository
-		logger: Logger
-		nextProject: NextJSProject
-	},
-	unknown
-> = async (ctx) => {
-	const response = await consola.prompt(
-		"Do you want to update your <Link>s for localised routing? (recommended)\nThis will replace any imports from next/link and next/navigation with their localised counterparts",
-		{
-			type: "confirm",
-			initial: true,
-		}
-	)
-
-	if (!response) return ctx
-	return await SetUpI18nRoutingFlow({ ...ctx, srcRoot: ctx.nextProject.srcRoot })
-}
